Refactor: Todo Page 코드 수정
